requirements.txt
numpy==2.1.0
Pillow==10.4.0
scikit_learn==1.4.2
torch==2.3.0+cu118
torchvision==0.18.0+cu118
